+++ src/roles/RoleInstances/harvester.ts
@@ -2,12 +2,20 @@ import { RoleName } from "enums/RoleName";
 import { Tasks } from "../creep-tasks/Tasks";
 import { Role } from "./Role";
 
-export class RoleRepairer extends Role {
-    public static roleName: RoleName = RoleName.Repairer;
+export class RoleHarvester extends Role {
+    public static roleName: string = RoleName.Harvester;
 
     public newTask(creep: Creep): void {
         if (creep.store.energy < creep.store.getCapacity()) {
-            // Find Structures for getting energy
+            // find the active source with the least harvesters currently on it
+            const sources = creep.room.find(FIND_SOURCES_ACTIVE);
+            if (sources != null) {
+                const leastBusySource = _.sortBy(sources, source => source.targetedBy.length)[0];
+                creep.task = Tasks.harvest(leastBusySource);
+                return;
+            }
+
+            // find other energy sources if no in game sources are active
             const structures = creep.room.find(FIND_STRUCTURES);
 
             // Pull energy from storage
@@ -17,22 +25,56 @@ export class RoleRepairer extends Role {
                 return;
             }
 
+            // Pickup dropped energy
+            const droppedEnergy = creep.pos.findClosestByPath(FIND_DROPPED_RESOURCES, { filter: (d) => d.resourceType === RESOURCE_ENERGY });
+            if (droppedEnergy != null) {
+                creep.task = Tasks.pickup(droppedEnergy);
+                return;
+            }
+
             // Withdraw from containers
             const containersWithEnergy = structures.filter(s => s.structureType === STRUCTURE_CONTAINER && s.store[RESOURCE_ENERGY] > 0) as StructureContainer[];
             if (containersWithEnergy.length > 0) {
                 creep.task = Tasks.withdraw(containersWithEnergy[0], RESOURCE_ENERGY);
                 return;
             }
+        }
 
-            // Resort to harvesting if no containers or storage
-            const sources = creep.room.find(FIND_SOURCES_ACTIVE);
-            const leastBusySource = _.sortBy(sources, source => source.targetedBy.length)[0];
-            creep.task = Tasks.harvest(leastBusySource);
+        // Look for spawns that aren't full first
+        const spawn = creep.pos.findClosestByPath(FIND_MY_SPAWNS, {
+            filter: s => s.store[RESOURCE_ENERGY] < s.store.getCapacity(RESOURCE_ENERGY)
+        });
+        if (spawn) {
+            creep.task = Tasks.transfer(spawn);
             return;
         }
 
+        // Look for extensions that aren't full next
+        const extension = creep.pos.findClosestByPath(FIND_STRUCTURES, {
+            filter: s => s.structureType === STRUCTURE_EXTENSION && s.store[RESOURCE_ENERGY] < s.store.getCapacity(RESOURCE_ENERGY)
+        });
+        if (extension && extension.structureType === STRUCTURE_EXTENSION) {
+            creep.task = Tasks.transfer(extension);
+            return;
+        }
 
-        // Repair main structures
+        // Look for tower that isn't full
+        const tower = creep.pos.findClosestByPath(FIND_MY_STRUCTURES, {
+            filter: s => s.structureType === STRUCTURE_TOWER && s.store[RESOURCE_ENERGY] < s.store.getCapacity(RESOURCE_ENERGY)
+        });
+        if (tower && tower.structureType === STRUCTURE_TOWER) {
+            creep.task = Tasks.transfer(tower);
+            return;
+        }
+
+        // Build instead
+        const sites = creep.room.find(FIND_MY_CONSTRUCTION_SITES);
+        if (sites.length) {
+            creep.task = Tasks.build(sites[0]);
+            return;
+        }
+
+        // Repair main structures instead
         const mainRepairs = creep.pos.findClosestByPath(FIND_STRUCTURES, {
             filter: s => {
                 return (s.structureType === STRUCTURE_EXTENSION ||
@@ -57,30 +99,7 @@ export class RoleRepairer extends Role {
             return;
         }
 
-        // Repair roads
-        var roadRepairs = creep.pos.findClosestByRange(FIND_STRUCTURES, {
-            filter: (structure) => {
-                return (structure.structureType === STRUCTURE_ROAD) && structure.hits < structure.hitsMax;
-            }
-        });
-        if (roadRepairs != null) {
-            creep.task = Tasks.repair(roadRepairs);
-            return;
-        }
-
-        // Repair walls
-        const wallRepairs = creep.pos.findClosestByRange(FIND_STRUCTURES, {
-            filter: (s) => {
-                return (s.structureType === STRUCTURE_WALL ||
-                    s.structureType === STRUCTURE_RAMPART) && s.hits <= 20000;
-            }
-        });
-        if (wallRepairs != null) {
-            creep.task = Tasks.repair(wallRepairs);
-            return;
-        }
-
-        // Upgrade if nothing left to repair
+        // Upgrade instead
         if (creep.room.controller) {
             creep.task = Tasks.upgrade(creep.room.controller);
             return;
