+++ src/roles/RoleInstances/upgrader.ts
@@ -1,6 +1,6 @@
 import { RoleName } from "enums/RoleName";
-import { Tasks } from "../creep-tasks/Tasks";
-import { Role } from "./Role";
+import { Tasks } from "../../creep-tasks/Tasks";
+import { Role } from "../Role";
 
 export class RoleUpgrader extends Role {
     public static roleName: RoleName = RoleName.Upgrader;
@@ -11,14 +11,15 @@ export class RoleUpgrader extends Role {
             const structures = creep.room.find(FIND_STRUCTURES);
 
             // Pull energy from storage
-            const storageWithEnergy = structures.filter(s => s.structureType === STRUCTURE_STORAGE && s.store[RESOURCE_ENERGY] > 0) as StructureStorage[];
-            if (storageWithEnergy.length > 0) {
-                creep.task = Tasks.withdraw(storageWithEnergy[0], RESOURCE_ENERGY);
+            if (creep.room.storage && creep.room.storage.store.getUsedCapacity(RESOURCE_ENERGY) > 0) {
+                creep.task = Tasks.withdraw(creep.room.storage, RESOURCE_ENERGY);
                 return;
             }
 
             // Withdraw from containers
-            const containersWithEnergy = structures.filter(s => s.structureType === STRUCTURE_CONTAINER && s.store[RESOURCE_ENERGY] > 0) as StructureContainer[];
+            const containersWithEnergy = creep.room.find(FIND_STRUCTURES, {
+                filter: s => s.structureType === STRUCTURE_CONTAINER && s.store.getUsedCapacity(RESOURCE_ENERGY) > 0
+            }) as StructureContainer[];
             if (containersWithEnergy.length > 0) {
                 creep.task = Tasks.withdraw(containersWithEnergy[0], RESOURCE_ENERGY);
                 return;
