+++ src/roles/index.ts
@@ -1,11 +1,11 @@
 import { RoleName } from "enums/RoleName";
-import { RoleBuilder } from "./builder";
-import { RoleHarvester } from "./harvester";
-import { RoleUpgrader } from "./upgrader";
-import { RoleContainerMiner } from "./containerMiner";
-import { RoleRepairer } from "./repairer";
-import { RoleTransporter } from "./transporter";
 import { Role } from "./Role";
+import { RoleBuilder } from "./RoleInstances/builder";
+import { RoleContainerMiner } from "./RoleInstances/containerMiner";
+import { RoleHarvester } from "./RoleInstances/harvester";
+import { RoleRepairer } from "./RoleInstances/repairer";
+import { RoleTransporter } from "./RoleInstances/transporter";
+import { RoleUpgrader } from "./RoleInstances/upgrader";
 
 export const roleDictionary: { [name in RoleName]: Role } = {
     [RoleName.Harvester]: new RoleHarvester(),
